# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FragEngine CS2 Data API",
    "description": "Professional-grade CS2 performance data, team statistics, and player metrics. Now with 4 flexible tiers starting from $0/mo.",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "https://cs2-datapipeline.onrender.com",
      "description": "Production Server"
    }
  ],
  "paths": {
    "/teams": {
      "get": {
        "summary": "Get all teams [BASIC]",
        "description": "Returns a list of all professional teams.",
        "responses": {
          "200": { "description": "Success" }
        }
      }
    },
    "/teams/{id}": {
      "get": {
        "summary": "Get team details [BASIC]",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Success" }
        }
      }
    },
    "/teams/{id}/players": {
      "get": {
        "summary": "Get team roster [STARTER]",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Success" },
          "403": { "description": "Tier Upgrade Required (STARTER+)" }
        }
      }
    },
    "/teams/{id}/map-stats": {
      "get": {
        "summary": "Get team map performance [PRO]",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Success" },
          "403": { "description": "Tier Upgrade Required (PRO+)" }
        }
      }
    },
    "/players/{id}/stats": {
      "get": {
        "summary": "Get granular player metrics [ULTRA]",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": { "description": "Success" },
          "403": { "description": "Tier Upgrade Required (ULTRA)" }
        }
      }
    },
    "/h2h/{team1}/{team2}": {
      "get": {
        "summary": "Historical H2H search [PRO]",
        "parameters": [
          {
            "name": "team1",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "team2",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Success" },
          "403": { "description": "Tier Upgrade Required (PRO+)" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RapidAPI-Key"
      }
    }
  },
  "security": [
    { "ApiKeyAuth": [] }
  ]
}